refactor(test): refactor testsuite and improve error handling
