--- cmd/al-ctl/cmd/test/utils.go
@@ -1,53 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- */
-package test
-
-import (
-	"fmt"
-	"gopkg.in/yaml.v3"
-	"os"
-	"path/filepath"
-)
-
-func getBaseSchemaVersionFromExtendedSchema() (string, error) {
-	file, err := os.Open(schemaPath)
-	if err != nil {
-		fmt.Println(err)
-		return "", err
-	}
-	defer file.Close()
-
-	decoder := yaml.NewDecoder(file)
-	var data map[string]interface{}
-
-	if err := decoder.Decode(&data); err != nil {
-		fmt.Println(err)
-		return "", err
-	}
-	imports, ok := data["imports"].([]interface{})
-	if !ok {
-		return "", fmt.Errorf("imports not found in extended schema")
-	}
-	if len(imports) < 2 {
-		return "", fmt.Errorf("reference to base schema not found in extended schema")
-	}
-	baseSchemaVersion := imports[1].(string)
-	return baseSchemaVersion, nil
-}
-
-func addVolumeInService(service *Service, currentDir string, pathInHost string, volume string) {
-	volume = filepath.Join(currentDir, pathInHost) + fmt.Sprintf(":/app/src/cdm/%s", volume)
-	service.Volumes = append(service.Volumes, volume)
-}
-
-func addSchemaEntrypointInService(service *Service, schemaFileName string) {
-	service.Entrypoint = append(service.Entrypoint, "-s", fmt.Sprintf("/app/src/cdm/%s", schemaFileName))
-}
-
-func addAssetEntrypointInService(service *Service, assetFileName string) {
-	service.Entrypoint = append(service.Entrypoint, "/app/src/cdm/"+assetFileName)
-}
